rerfact: IGameBoardComponent삭제, 모든 컴포넌트가 IGameComponent만을 상속받도록 변경
