+++ src/tetrisgame/TetrisGame.java
@@ -8,18 +8,27 @@ import gamestarter.GameStarter;
 import graphics.screens.GameScreen;
 import tetrisgame.enumerations.eDifficulty;
 import tetrisgame.enumerations.eMsg;
-import tetrisgame.parts.component.Tetromino;
+import tetrisgame.parts.Position;
 import tetrisgame.parts.Timer;
-import tetrisgame.parts.component.GameBoard;
+import tetrisgame.parts.component.RemovingAnim;
 import tetrisgame.parts.component.Score;
+import tetrisgame.parts.component.Tetromino;
+import tetrisgame.parts.component.Tile;
 
 public class TetrisGame implements Runnable {
 	protected GameScreen mScreen;
 
 	private volatile boolean mbGameOverFlag;
 
-	private GameBoard mGameBoard;
+	public Tile mBoard[][];
+	private Tetromino mCurrTetromino;
+	private Tetromino mNextTetromino;
+	private RemovingAnim mRemoveAnim;
 	private Score mScore;
+
+	private boolean mbRemovingFlag;
+	private int mRemoveColArr[];
+	private int mNumRemovableLines;
 
 	private Timer mTimer;
 
@@ -28,12 +37,15 @@ public class TetrisGame implements Runnable {
 	private volatile boolean mbKeyReactFlag;
 	private static final float KEY_REACT_TIME = 0.1f;
 
+	public static final int BOARD_COL = 20;
+	public static final int BOARD_ROW = 10;
+
 	public TetrisGame(GameScreen gameScreen) {
 		mScreen = gameScreen;
 
-		mGameBoard = new GameBoard(this);
-		mGameBoard.clear();
-
+		mBoard = new Tile[BOARD_COL][BOARD_ROW];
+		mCurrTetromino = new Tetromino(this, mBoard);
+		mNextTetromino = new Tetromino(this, mBoard);
 		mScore = new Score(this);
 
 		// TODO:
@@ -42,6 +54,8 @@ public class TetrisGame implements Runnable {
 
 		mTimer = new Timer();
 		mTimer.initialize();
+
+		mRemoveColArr = new int[4];
 
 		mCurrKeyCode = -1;
 		mKeyReactTimeTick = 0.f;
@@ -53,8 +67,67 @@ public class TetrisGame implements Runnable {
 			mbGameOverFlag = true;
 			return;
 		}
-		mGameBoard.react(msg);
+		if (msg == eMsg.COLL_WITH_FLOOR) {
+			boolean bRemovable = true;
+			int posCol = mCurrTetromino.getPosition().mCol;
+			int startCol = posCol + 3 < BOARD_COL ? posCol + 3 : BOARD_COL - 1;
+			posCol = posCol >= 0 ? posCol : 0;
+			for (int boardCol = startCol; boardCol >= posCol; boardCol--) {
+				for (Tile currBolck : mBoard[boardCol]) {
+					if (currBolck == null) {
+						bRemovable = false;
+						break;
+					}
+				}
+				if (bRemovable) {
+					mRemoveColArr[mNumRemovableLines++] = boardCol;
+				}
+				bRemovable = true;
+			}
+			if (mNumRemovableLines > 0) {
+				mCurrTetromino.react(msg);
+				mRemoveAnim = new RemovingAnim(this, mBoard, mRemoveColArr, mNumRemovableLines);
+				mbRemovingFlag = true;
+				for (int j = 0; j < mNumRemovableLines; j++) {
+					this.broadcast(eMsg.LINE_REMOVE);
+				}
+			} else {
+				mCurrTetromino = mNextTetromino;
+				mNextTetromino = new Tetromino(this, mBoard);
+				if (checkGameOver()) {
+					this.broadcast(eMsg.GAME_OVER);
+				}
+			}
+		}
+		if (msg == eMsg.REMOVE_ANIM_OVER) {
+			mCurrTetromino = mNextTetromino;
+			mNextTetromino = new Tetromino(this, mBoard);
+			mbRemovingFlag = false;
+			mRemoveAnim = null;
+
+			if (checkGameOver()) {
+				this.broadcast(eMsg.GAME_OVER);
+			}
+
+			removeLines();
+			fallDownLines();
+			mNumRemovableLines = 0;
+		}
+
+		if (mbRemovingFlag) {
+			mCurrTetromino.react(msg);
+		}
+		if (mRemoveAnim != null) {
+			mRemoveAnim.react(msg);
+		}
 		mScore.react(msg);
+		for (Tile tileLine[] : mBoard) {
+			for (Tile tile : tileLine) {
+				if (tile != null) {
+					tile.react(msg);
+				}
+			}
+		}
 	}
 
 	@Override
@@ -65,7 +138,6 @@ public class TetrisGame implements Runnable {
 		}
 		// Save user record
 		String userName = JOptionPane.showInputDialog("Enter your name");
-		System.out.println(userName);
 		GameStarter.setOver();
 	}
 
@@ -86,7 +158,19 @@ public class TetrisGame implements Runnable {
 			mKeyReactTimeTick = 0.f;
 		}
 
-		mGameBoard.update(deltaTime, userInput);
+		if (mbRemovingFlag == false) {
+			mCurrTetromino.update(deltaTime, userInput);
+		}
+		if (mRemoveAnim != null) {
+			mRemoveAnim.update(deltaTime, userInput);
+		}
+		for (Tile tileLine[] : mBoard) {
+			for (Tile tile : tileLine) {
+				if (tile != null) {
+					tile.update(deltaTime, userInput);
+				}
+			}
+		}
 	}
 
 	public synchronized void getUserInput(KeyEvent e) {
@@ -119,8 +203,16 @@ public class TetrisGame implements Runnable {
 		}
 	}
 
-	public GameBoard getGameBoard() {
-		return mGameBoard;
+	public Tile[][] getGameBoard() {
+		return mBoard;
+	}
+
+	public Tetromino getCurrTetromino() {
+		return mCurrTetromino;
+	}
+
+	public Tetromino getNextTetromion() {
+		return mNextTetromino;
 	}
 
 	public long getCurrScore() {
@@ -130,5 +222,53 @@ public class TetrisGame implements Runnable {
 	private void draw() {
 		// TODO: refact
 		mScreen.repaint();
+	}
+
+	private boolean checkGameOver() {
+		boolean collResult = collisionTest();
+		return collResult;
+	}
+
+	public void removeLines() {
+		for (int index = 0; index < mNumRemovableLines; index++) {
+			int col = mRemoveColArr[index];
+			for (int row = 0; row < BOARD_ROW; row++) {
+				mBoard[col][row] = null;
+			}
+		}
+	}
+
+	public void fallDownLines() {
+		for (int index = 0; index < mNumRemovableLines; index++) {
+			int col = mRemoveColArr[index];
+			Tile removedLine[] = mBoard[col];
+			for (int c = col; c > 0; c--) {
+				mBoard[c] = mBoard[c - 1];
+			}
+			for (int i = index; i < mNumRemovableLines; i++) {
+				mRemoveColArr[i]++;
+			}
+			mBoard[0] = removedLine;
+		}
+	}
+
+	private boolean collisionTest() {
+		boolean re = false;
+		Position pos = mCurrTetromino.getPosition();
+		for (int c = 0; c < Tetromino.SHAPE_COL; c++) {
+			for (int r = 0; r < Tetromino.SHAPE_ROW; r++) {
+				if (mCurrTetromino.mShape[c][r] != null) {
+					int col = pos.mCol + c;
+					int row = pos.mRow + r;
+					if (BOARD_COL <= col || col < 0 || row < 0 || BOARD_ROW <= row ||
+							mBoard[col][row] != null) {
+						re = true;
+						return re;
+					}
+
+				}
+			}
+		}
+		return re;
 	}
 }
