--- src/tetrisgame/enumerations/eCollResult.java
@@ -1,6 +1,0 @@
-package tetrisgame.enumerations;
-
-public enum eCollResult {
-    COLL,
-    NOT_COLL
-}
